Day 1 2 3 setup with initial assets done
For the most part finished setting up each of the three days with all the available assets. Also installed an FBX exporter so anyone can export anything from the scene. Re modelled the building walls to include window hole and separated the walls in the storage room so they can be textured separately from the rest of the building.