MANY CHANGES!!!!
Added in the lockers, a new downlight prop, new lightbulb, new cathouse (with litter included). I also went back through most of the assets in substance painter and baked ambient occlusion into all of their albedo maps. This means they'll look WAYYYYY better than they currently do. I didn't do it for a few props, namely the cathouse and couch, since baking AO isn't a very good idea for props that have moving parts (the doors on the cathouse being the issue, and with the couch, since I only modelled one cushion and pillow, the AO would look completely wrong when I duplicated the model a bunch of times to fill the rest of the couch.) Hopefully this doesn't cause a bunch of merge conflicts. We shall see.