Fixed Cat Rig!!!!!!!!!!!!!!!!!!
Finally managed to stop the cat rig looking like shit whenever I moved any of the bones. Turns out there was a skin quality setting hiding in the project settings that was overriding all the skin quality changes I was making on the model itself. Anyway, this is good news for everyone since we can pose the cats however we like in Unity.  No textures yet (in da pipeline).