Added in new props to day 1
Added in the new props to scene from the merge. We now have a desert cabinet, some fans, cupboards, mop sink. Also changed the ground plane texture to be a white radial gradient. Makes the view outside look a little more pleasant. The cat model is also now in the game and can be posed (it's a takes a lot of fiddling to get a pose that looks good, so I left it out of the scene for now).