Added dust particles, adjusted day 1+3 lightmaps
The dust particles are in the scene now under each light, as well as a global particle system which is temporary (will change the sprites to be things like cat hair). Plus, a bunch of small changes the materials to make them not look odd with reflections. I haven't touched day 2 yet since it takes a while to set up each light and I think it'd be faster to work out the best settings in day 1 and 3 then just copy it over to day 2.