Fixed big lighting bug on all days
Essentially, any prefab that wasn't unpacked was trying to use lightmaps from different scenes, this lead to some odd lighting issues. Therefore, I unpacked every single prefab in each scene (minus the cat jails in day 3, please don't unpack those.) The issue with doing that is that you can't just edit one prefab and apply it to all other prefabs, you need to change each object individually. I think we're past the point of making changes like that though, so it was worth the trade off. I also apologise for bombarding the main branch with commits, but it must be done.