Made props appear in players hand for each day
Very basic implementation. Essentially just sets the object to be active or inactive in the task outcomes script. The placements of the props look a little funny in day 2 and 3, but that's just because they don't have any animations yet. Also, I'm not done yet. I still need to set up the object position highlighter, mkay.