Implemented final big red button task
Plays the animation of all the jail cell doors opening, although the scene ends too fast, so that's something that should be adjusted. I also changed the poop model since the lighting on it was well and truly cooked for whatever reason, and the only way I could fix it was to just make a new model. Didn't get around to having the cups and poop appear in front of the player. I thought about it and decided I'd rather just finish the ending, since it was a tad more important for the beta.