Removed cup collecting task from day 3
From now on, when the player is prompted to go check the taskboard after scooping the poop, interacting with the taskboard will teleport them into the ending hallway. The reason I changed this was that it was more unexpected and far easier to force the player into the staffroom, and also because doing the same tasks 3 times over was getting a bit bland.