Made all the tasks represent Bob's thoughts, added particles into every day
I made a a lot of changes to the tasks descriptions so that they represented what Bob was thinking. It was difficult to tell the player what to do whilst also being what Bob would be think, so I think there's room for improvement. The particles on day 3 are now just strands of hair.