Added in basic water refraction shader. It's raining on day 2 now as well.
Day 1 has the refraction shader to make the skyline warp and wobble a bit, implying there's water. Day 2 I used the same shader to make the window look wet, and also added a basic rain particle system that just uses trails to create the rain drops. On day 3 I repurposed the shader again, this time using it on the window because it looked weird.