Hair particles now play when hair is vacuumed
I did it in a separate script since it was easier. Essentially, when the hair is destroyed, the hair particle system (which is always in the scene) is moved to the position of the hair and played once. Simple. I don't think it would make sense to have particles for the poop or cups, so I don't think there's any more work needed here.