Adjusted Clickable Object script so that it's possible to change crosshair on hover
What I'm doing now is checking if the object is on screen, then checking how far from the player it is. The reason I did it like this was so that I didn't need to put a trigger collider on every interactable object (like the task system). I tried to optimize it a tad to stop the raycasts happening at ALL times, so it should only raycast when you're close enough. Hopefully this doesn't kill the framerate (it seemed fine when I did a quick test)