Added task and dialogue UI images (not final)
Still a lot that needs to be done with them, namely making sure the task UI isn't on until after the phone call, or during the ending hallway sequence. Nonetheless, it's in the game and it works well enough. I also got rid of that error message related to the hair particles script that I wrote. The reason the UI was appearing in front of the text was because the text needs to be below the background in the hierarchy.