Added a van to the outside of day 1
Modelling cars that don't look like shit is far harder than I imagined, so I went with a van, which is far easier to make. Right now it's only in day 1 and has a super basic drive animation that goes up the road one way, but I think it's fine for now.