Crosshair changes on clickable objects now (not for tasks yet)
Doing it for the tasks is relatively easy, the issue is that I need to do a check for the currently active task before turning the interaction crosshair on, and I wasn't sure how to do that. If I don't do that, the crosshair will change on every single task object, regardless of whether the associated task is active or not.