Task List on taskboard changes per task now (plus more!)
I've kinda changed the purpose the taskboard to make more sense. Basically, the reason you keep going back to the taskboard is so that you can tick off the task you just did alongside seeing the next task. The reason Bob keeps doing this is because he's paranoid about the boss watching him (which is true, there's security cameras in every room). I also did some lightmap rubbish. There's one wall in the staffroom that's causing me a bit of grief, mainly due to the fact that it just has a bunch of weird artifacts on it for no reason. Besides that one issue, the lighting is basically done. I've tried matching the lighting in the ending room a little better so the transition isn't as obvious, and it's improved somewhat, but I still need to tweak that. Lastly, I changed some of the task descriptions to show Bob becoming a tad weirder on day 2. Let me know if anything has broken, I didn't do a complete playthrough to check.