Changed the clickable objects script to use a trigger instead
I was very adamant on not using a trigger to play the animation for some reason, but upon further testing I see that it is by far the best way to go about doing this. The thing that made me change it was seeing how the dessert fridge door worked without the trigger system (it wasn't able to stop in one state, it would always have to go back to the start of the animation, which sucked since you couldn't leave the door open.) Anyway, the new system is pretty similar, except now with a trigger parameter in the animation controller called Interaction (you can name the animations whatever you want now, hooray) If you wanna see how it works, just click on the cat yarn ball or fridge door in day 1. I also fixed some lighting rubbish no one needs to know about.