Added the ability to play particles on click for interactable objects (coffee machine can be clicked on now)
Also had to make a new script for changing the hover icon, since I needed to have only one script control it as opposed to a bunch of scripts. It isn't perfect, but it's good enough. All you need to do to make it work is plug in the particle system you want to play into to script on the object. Look at the coffee machine in day 1 if you need to know how it works.