Added ability to pick up pills anytime before flipping sign (read desc)
Basically, since I noticed some people during playtesting were trying to pick up the antihistamines BEFORE answering the phone or flipping the sign, I added in the option that would allow them to do that. What happens now is that the task telling the player to pick up the antihistamines will only activate if the player hasn't already picked them up. If they haven't, it'll just skip over it. I think it's pretty neat, and worth the effort to include, since it's a logical thing to be able to do in the game. A new problem has arisen though. Some tasks for god knows what reason will not activate the hover icon. I tried and tried and tried to figure out why, but I just can't do it. I feel like it's something super simple though, maybe related to colliders, since it seems to happen on random tasks. Anyway, essay concluded.