Gave interaction and task hover icons separate sprites
I also spent like 3 hours banging my head against the wall trying to figure out how to stop other task objects interactable colliders from messing with the ray cast. I'm fairly certain the reason this is happening because I'm doing the icon change in the interactable script, which is on multiple objects running at the same time, which means the hover icon doesn't know which collider to use. My solution I had in mind was to disable all the interaction colliders on task objects that aren't active, but I did this (it's still in the code, just commented out) and it broke the hover icon entirely.