Completely removed the interaction collider stuff for reasons(read desc)
I mentioned in a previous commit that the interaction colliders were causing me a lot of trouble when trying to get the task hover icon working. I came to the conclusion that it would be much easier to just remove the colliders altogether, and adjust the code so that it didn't require them. I knew this might have an impact on performance since the player will always be raycasting (probably multiple times a frame as well given there's multiple interactable scripts running), but my frame rate seemed more than okay when testing. This isn't exactly conclusive evidence that all is well, so I'll test it a bit more. The other drawback to removing the colliders is that the player can interact with task objects from any position (assuming they're within the 1.5m ray length). I don't consider this a huge issue, since there's only like two tasks where you can complete them through walls. An easy way to control this would be to have an empty object on the interactable layer with a collider that would block the ray from continuing. It's a bit hacky, but this is the last week before submission, so as long as it works and doesn't kill the framerate, I think it's fine.