Fixed the weirdness with day 1 vacuum pickup (hopefully)
I honestly have no idea why this is acting up. The actual picking up of the object is working exactly as it should, it's just the hover icon that's being activated at weird times. My solution was to move the vacuum so that its pivot wasn't so far from the actual model (that shouldn't have any impact, but it did) and also making the antihistamines box collider smaller so there was less chance of it clashing. Testing this, it still isn't doing what I want it to do exactly, but it's close enough.