Added new interactions (read description)
A few were on day 3. These were the salt shaker, the flying forks and the boss picture frame that you can throw on the floor. On day 1 I also added the ability to close the cat cage doors, so if the player wants to do that they can. The cafe logo is also on the front door now. Also, the weirdness with the day 1 vacuum is not fixed. I have no idea why, and it's making me sad. It works well enough, but I there's just something slightly off with it that makes the hover icon act funny. I've even tried just copy and pasting the day 2 vacuum, but the problem still exists, so it doesn't seem to be a problem with the vacuum itself. I'll keep trying, maybe I'll work it out.