Fixed the weird vacuum pickup UI bug (for real this time)
There actually wasn't anything wrong with my code. The issue was that there was a singular object in the scene that wasn't on the Interactable layer, and in the script the raycast works by only hitting the layer of the object with the interactable script. As a result, the object that wasn't on the right layer was messing with the raycast and causing it to hit the wrong things. I'm not lying when I say I spent atleast 4 hours trying to fix this bug, and the solution was literally just two mouse clicks.