Added in all the awful cat sounds for the ending hallway.
The ending is kind of disturbing now. I love it. I also fixed all those weird shimmering effects that happened around really bright objects. It was happening due to the fact that the material emission was set way too high (in order to meet the bloom threshold). I fixed it by lowering the threshold massively and then just going through each material with emission and lowering the values till it looked right. There's still a little shimmer when you look at the street lights outside, but that's more to do with the window shader.