Adjusted the laser particles a tiny bit
Due to the way I've set up the particles that play when you destroy the poop, the poop will be destroyed before the particles hit it, which looks odd. Since I couldn't be bothered spending a bunch of time redoing the way I call the particles, I just made the particles appear a bit quicker. The poop still disappears before the laser hits it, but it's better than before.