Changed menu buttons code
Basically made it so that clicking the controls would hide the instructions, and the other way round. Before I changed that, the credits would always appear above the controls, and you'd have to turn off the credits before you could see them. Still haven't managed to work out why that stutter is happening at the end of the game when loading the menu.