End game menu stutter gone, lightmaps compressed, intro music realigned
The lightmap compression didn't actually reduce the file size all that much (~200 MB removed), but it should help with the load times. I also turned on mip maps for the lightmaps so that they aren't constantly rendering at 100% quality regardless of distance. To realign the intro music I just shortened the actual audio file (there was a few seconds of silence at the beginning that I cut down). I'd say at this point we could put this on Itch.